@robodoo 2025-03-11
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/201243.png)](https://mergebot.odoo.com/odoo/odoo/pull/201243)