@adda-odoo 2025-03-21
hi @odoo/rd-logistics can this get a quick review? Thanks!