@clesgow 2025-03-28
Hello @adda-odoo, as the retargeted PR has been merged, can we close this one?